# pkg/api/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  AuthConfig:
    properties:
      password:
        type: string
        x-go-name: Password
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
...
  Duration:
    description: |-
      A Duration represents the elapsed time between two instants
      as an int64 nanosecond count. The representation limits the
      largest representable duration to approximately 290 years.
    format: int64
    type: integer
    x-go-package: time
  ForwarderConfig:
    properties:
      selector:
        $ref: '#/definitions/SelectorConfig'
      targets:
        items:
          type: string
        type: array
        x-go-name: Targets
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  HandlerConfig:
    properties:
      auth:
        $ref: '#/definitions/AuthConfig'
      auther:
        type: string
        x-go-name: Auther
      chain:
        type: string
        x-go-name: Chain
      metadata:
        additionalProperties:
          type: object
        type: object
        x-go-name: Metadata
      retries:
        format: int64
        type: integer
        x-go-name: Retries
      tls:
        $ref: '#/definitions/TLSConfig'
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
...
  ListenerConfig:
    properties:
      auth:
        $ref: '#/definitions/AuthConfig'
      auther:
        type: string
        x-go-name: Auther
      chain:
        type: string
        x-go-name: Chain
      metadata:
        additionalProperties:
          type: object
        type: object
        x-go-name: Metadata
      tls:
        $ref: '#/definitions/TLSConfig'
      type:
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
...
  NameserverConfig:
    properties:
      addr:
        type: string
        x-go-name: Addr
      chain:
        type: string
        x-go-name: Chain
      clientIP:
        type: string
        x-go-name: ClientIP
      hostname:
        type: string
        x-go-name: Hostname
      prefer:
        type: string
        x-go-name: Prefer
      timeout:
        $ref: '#/definitions/Duration'
      ttl:
        $ref: '#/definitions/Duration'
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
...
  ResolverConfig:
    properties:
      name:
        type: string
        x-go-name: Name
      nameservers:
        items:
          $ref: '#/definitions/NameserverConfig'
        type: array
        x-go-name: Nameservers
      type:
        description: inline, file, etc.
        type: string
        x-go-name: Type
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  Response:
    properties:
      code:
        format: int64
        type: integer
        x-go-name: Code
      msg:
        type: string
        x-go-name: Msg
    type: object
    x-go-package: github.com/go-gost/gost/pkg/api
  SelectorConfig:
    properties:
      failTimeout:
        $ref: '#/definitions/Duration'
      maxFails:
        format: int64
        type: integer
        x-go-name: MaxFails
      strategy:
        type: string
        x-go-name: Strategy
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  ServiceConfig:
    properties:
      addr:
        type: string
        x-go-name: Addr
      bypass:
        type: string
        x-go-name: Bypass
      forwarder:
        $ref: '#/definitions/ForwarderConfig'
      handler:
        $ref: '#/definitions/HandlerConfig'
      hosts:
        type: string
        x-go-name: Hosts
      listener:
        $ref: '#/definitions/ListenerConfig'
      name:
        type: string
        x-go-name: Name
      resolver:
        type: string
        x-go-name: Resolver
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
  TLSConfig:
    properties:
      caFile:
        type: string
        x-go-name: CAFile
      certFile:
        type: string
        x-go-name: CertFile
      keyFile:
        type: string
        x-go-name: KeyFile
      secure:
        type: boolean
        x-go-name: Secure
      serverName:
        type: string
        x-go-name: ServerName
    type: object
    x-go-package: github.com/go-gost/gost/pkg/config
info:
  title: Documentation of Web API.
  version: 1.0.0
paths:
...
  /config/resolvers:
    post:
      operationId: createResolverRequest
      parameters:
      - in: body
        name: data
        schema:
          $ref: '#/definitions/ResolverConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/createResolverResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Create a new resolver, the name of the resolver must be unique in resolver
        list.
      tags:
      - ConfigManagement
  /config/resolvers/{resolver}:
    delete:
      operationId: deleteResolverRequest
      parameters:
      - in: path
        name: resolver
        required: true
        type: string
        x-go-name: Resolver
      responses:
        "200":
          $ref: '#/responses/deleteResolverResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Delete resolver by name.
      tags:
      - ConfigManagement
    put:
      operationId: updateResolverRequest
      parameters:
      - in: path
        name: resolver
        required: true
        type: string
        x-go-name: Resolver
      - in: body
        name: data
        schema:
          $ref: '#/definitions/ResolverConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/updateResolverResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Update resolver by name, the resolver must already exist.
      tags:
      - ConfigManagement
  /config/services:
    post:
      operationId: createServiceRequest
      parameters:
      - in: body
        name: data
        schema:
          $ref: '#/definitions/ServiceConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/createServiceResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Create a new service, the name of the service must be unique in service
        list.
      tags:
      - ConfigManagement
  /config/services/{service}:
    delete:
      operationId: deleteServiceRequest
      parameters:
      - in: path
        name: service
        required: true
        type: string
        x-go-name: Service
      responses:
        "200":
          $ref: '#/responses/deleteServiceResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Delete service by name.
      tags:
      - ConfigManagement
    put:
      operationId: updateServiceRequest
      parameters:
      - in: path
        name: service
        required: true
        type: string
        x-go-name: Service
      - in: body
        name: data
        schema:
          $ref: '#/definitions/ServiceConfig'
        x-go-name: Data
      responses:
        "200":
          $ref: '#/responses/updateServiceResponse'
      security:
      - basicAuth:
        - '[]'
      summary: Update service by name, the service must already exist.
      tags:
      - ConfigManagement
produces:
- application/json
responses:
...
  createResolverResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  createServiceResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
...
  deleteResolverResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  deleteServiceResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
...
  updateResolverResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
  updateServiceResponse:
    description: successful operation.
    headers:
      Data: {}
    schema:
      $ref: '#/definitions/Response'
schemes:
- https
- http
securityDefinitions:
  basicAuth:
    type: basic
swagger: "2.0"